pyproject.toml
[tool.poetry]
name = "pitcher-so-ml"
version = "0.1.0"
description = "Machine Learning project for pitcher strikeout predictions"
authors = ["Arturo Burigo"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^1.5.0"
numpy = "^1.21.0"
scikit-learn = "^1.0.0"
python-dateutil = "^2.8.2"
requests = "^2.28.0"
beautifulsoup4 = "4.9.3"
matplotlib = "^3.5.0"
seaborn = "^0.12.0"
python-dotenv = "^1.0.0"
xgboost = "^2.0.0"
lightgbm = "^4.1.0"
lxml = "^4.9.0"
html5lib = "^1.1"
python-telegram-bot = "13.7"
APScheduler = "3.6.3"
pytz = "2021.1"
telebot = "*"
selenium = "^4.0.0"
webdriver-manager = "^4.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
flake8 = "^6.0.0"
mypy = "^1.0.0"

[tool.poetry.scripts]
scrape-betano = "src.scrapping.scrapper_betano_pitchers_odds:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
